fresh_harvest/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
boto3==1.35.39
botocore==1.35.39
Brotli==1.1.0
cement==2.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.3
crispy-bootstrap4==2024.1
cssselect2==0.7.0
Django==5.0.2
django-appconf==1.0.6
django-chartjs==2.3.0
django-crispy-forms==2.1
django-image-cropping==1.7
django-mathfilters==1.0.0
django-sass-processor==1.4
django-slick-reporting==1.3.0
django-storages==1.14.4
factory-boy==3.3.0
Faker==24.3.0
fonttools==4.50.0
frozenlist==1.4.1
gunicorn==21.2.0
html5lib==1.1
idna==3.6
image==1.5.33
jmespath==1.0.1
multidict==6.0.5
packaging==24.0
pathspec==0.10.1
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.9.0
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.3
semantic-version==2.8.5
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
tinycss2==1.2.1
twilio==8.13.0
tzdata==2023.4
urllib3==1.26.18
wcwidth==0.1.9
weasyprint==61.2
webencodings==0.5.1
yarl==1.9.4
zopfli==0.2.3
